Update Layout
Admin GUI

+++ app/Controller/SuppliersController.php
@@ -23,8 +23,13 @@ class SuppliersController extends AppController {
  * @return void
  */
 	public function index() {
-		$this->Supplier->recursive = 0;
-		$this->set('suppliers', $this->paginate());
+		if ($this->Session->check('userSS') && $this->Session->check('passSS')) {
+			$this->Supplier->recursive = 0;
+			$this->set('suppliers', $this->paginate());
+		}
+		else{
+      		$this->redirect(array('controller'=>'users','action'=>'login'));
+    	}
 	}
 
 /**
@@ -35,11 +40,16 @@ class SuppliersController extends AppController {
  * @return void
  */
 	public function view($id = null) {
-		if (!$this->Supplier->exists($id)) {
-			throw new NotFoundException(__('Invalid supplier'));
+		if ($this->Session->check('userSS') && $this->Session->check('passSS')) {
+			if (!$this->Supplier->exists($id)) {
+				throw new NotFoundException(__('Invalid supplier'));
+			}
+			$options = array('conditions' => array('Supplier.' . $this->Supplier->primaryKey => $id));
+			$this->set('supplier', $this->Supplier->find('first', $options));
 		}
-		$options = array('conditions' => array('Supplier.' . $this->Supplier->primaryKey => $id));
-		$this->set('supplier', $this->Supplier->find('first', $options));
+		else{
+      		$this->redirect(array('controller'=>'users','action'=>'login'));
+    	}
 	}
 
 /**
@@ -48,15 +58,20 @@ class SuppliersController extends AppController {
  * @return void
  */
 	public function add() {
-		if ($this->request->is('post')) {
-			$this->Supplier->create();
-			if ($this->Supplier->save($this->request->data)) {
-				$this->Session->setFlash(__('Đã thêm nhà cung cấp '.$this->request->data['Supplier']['nameSupplier']), 'flash/success');
-				$this->redirect(array('action' => 'index'));
-			} else {
-				$this->Session->setFlash(__('Không thể thêm nhà cung cấp này. Vui lòng kiểm tra lại.'), 'flash/error');
+		if ($this->Session->check('userSS') && $this->Session->check('passSS')) {
+			if ($this->request->is('post')) {
+				$this->Supplier->create();
+				if ($this->Supplier->save($this->request->data)) {
+					$this->Session->setFlash(__('Đã thêm nhà cung cấp '.$this->request->data['nameSupplier']), 'flash/success');
+					$this->redirect(array('action' => 'index'));
+				} else {
+					$this->Session->setFlash(__('Không thể thêm nhà cung cấp này. Vui lòng kiểm tra lại.'), 'flash/error');
+				}
 			}
 		}
+		else{
+      		$this->redirect(array('controller'=>'users','action'=>'login'));
+    	}
 	}
 
 /**
@@ -67,21 +82,26 @@ class SuppliersController extends AppController {
  * @return void
  */
 	public function edit($id = null) {
-        $this->Supplier->id = $id;
-		if (!$this->Supplier->exists($id)) {
-			throw new NotFoundException(__('Invalid supplier'));
-		}
-		if ($this->request->is('post') || $this->request->is('put')) {
-			if ($this->Supplier->save($this->request->data)) {
-				$this->Session->setFlash(__('Cập nhật thành công'), 'flash/success');
-				$this->redirect(array('action' => 'index'));
-			} else {
-				$this->Session->setFlash(__('Không thể cập nhật. Vui lòng thử lại.'), 'flash/error');
+		if ($this->Session->check('userSS') && $this->Session->check('passSS')) {
+	        $this->Supplier->id = $id;
+			if (!$this->Supplier->exists($id)) {
+				throw new NotFoundException(__('Invalid supplier'));
 			}
-		} else {
-			$options = array('conditions' => array('Supplier.' . $this->Supplier->primaryKey => $id));
-			$this->request->data = $this->Supplier->find('first', $options);
+			if ($this->request->is('post') || $this->request->is('put')) {
+				if ($this->Supplier->save($this->request->data)) {
+					$this->Session->setFlash(__('Cập nhật thành công'), 'flash/success');
+					$this->redirect(array('action' => 'index'));
+				} else {
+					$this->Session->setFlash(__('Không thể cập nhật. Vui lòng thử lại.'), 'flash/error');
+				}
+			} else {
+				$options = array('conditions' => array('Supplier.' . $this->Supplier->primaryKey => $id));
+				$this->request->data = $this->Supplier->find('first', $options);
+			}
 		}
+		else{
+      		$this->redirect(array('controller'=>'users','action'=>'login'));
+    	}
 	}
 
 /**
@@ -93,20 +113,27 @@ class SuppliersController extends AppController {
  * @return void
  */
 	public function delete($id = null) {
-		if (!$this->request->is('post')) {
-			throw new MethodNotAllowedException();
-		}
-		$this->Supplier->id = $id;
-		if (!$this->Supplier->exists()) {
-			throw new NotFoundException(__('Invalid supplier'));
-		}   
-          
-        $supplier = $this->Supplier->find('first', array('conditions' => array('Supplier.' . $this->Supplier->primaryKey => $id)));
-           
-		if ($this->Supplier->delete()) {
-			$this->Session->setFlash(__('Đã xóa nhà cung cấp '. $supplier['Supplier']['nameSupplier']), 'flash/success');
+		if ($this->Session->check('userSS') && $this->Session->check('passSS')) {
+			if (!$this->request->is('post')) {
+				throw new MethodNotAllowedException();
+			}
+			$this->Supplier->id = $id;
+			if (!$this->Supplier->exists()) {
+				throw new NotFoundException(__('Invalid supplier'));
+			}   
+	          
+	        $supplier = $this->Supplier->find('first', array('conditions' => array('Supplier.' . $this->Supplier->primaryKey => $id)));
+	           
+			if ($this->Supplier->delete()) {
+				$this->Session->setFlash(__('Đã xóa nhà cung cấp '. $supplier['Supplier']['nameSupplier']), 'flash/success');
+				$this->redirect(array('action' => 'index'));
+			}
+			$this->Session->setFlash(__('Không thể xóa nhà cung cấp '. $supplier['Supplier']['nameSupplier']), 'flash/error');
 			$this->redirect(array('action' => 'index'));
 		}
-		$this->Session->setFlash(__('Không thể xóa nhà cung cấp '. $supplier['Supplier']['nameSupplier']), 'flash/error');
-		$this->redirect(array('action' => 'index'));
-	}}
+		else{
+      		$this->redirect(array('controller'=>'users','action'=>'login'));
+    	}
+	}
+
+}
